Update README.md
The -c flag in the kubectl exec command specifies which container in a pod you want to execute the command against. This is necessary when a pod has multiple containers, and you need to indicate which container should run the command.

If the -c flag is not passed, the command is executed, however a messagem is popupd, and in automated scripts it invalidade the processo.

The adition of -c will not affect the manual usage and will prevent erros in some automation process.

Summary
Use -c when the pod has multiple containers.
It ensures you are executing the command in the correct container.
Helps avoid unexpected behavior or errors when Kubernetes defaults to the wrong container.

Signed-off-by: Anselmo Battisti <anselmobattisti@gmail.com>